refactor: move repository interfaces to domain/port package
Create domain/port/ with all infrastructure interfaces (EventSaver,
SessionSaver, etc). Usecase layer uses type aliases for backward
compatibility. Infrastructure now imports domain/port instead of
application/usecase, fixing the forbidden dependency direction.
